ai_agents/requirements.txt
streamlit
langchain
langchain-openai
langchain-core
langchain-community
langsmith
python-dotenv
watchdog
